feat: Add option to configure animation duration
Problem:
Text2Motion API now supports passing a seconds parameter to configure
the duration of the generated animation. We want to port this feature
to the blender integration.

Solution:
Added new option panel for passing in the seconds parameter.

Risk:
Any bugs in the extension code could cause the extension to stop
working. We will mitigate this risk by testing before releasing.

Testing Done:
Ran the extension locally from devcontainer build and verified the
following:
1. Can still set API key
2. License info is still displaying
3. Prompt still works
4. Can change name of action
5. can toggle root motion
6. Auto duration works
7. Seconds and Frames can both be set and generate expected length
8. Verified the parameter are properly bounded

Made a windows and linux build and repeated the above steps on a local
blender installation.