chore: remove unused, fix typo, update readme
Problem:
There are a couple of small miscellaneous issues in the repo:
1. Readme have missing table of content section
2. There are some typos in the codebase
3. Spell checker ignore could be added for some external library words
4. Ruff found unused variables that should be removed

Solution:
Addressed the issues listed above.

Also added a section in Readme to include guideline for the Blender
Extension review. And added ruff to dev requirements so we can run the
linter command.

Risk:
These are mostly dev environment and refactoring changes, not expected
to have any functional impact.

Testing Done:
1. Tested in devcontainer build and verified that I can still send
request to the server.
2. Tested the same with local build install.